[ui] Added Hero Component with responsiveness
