+++ packages/ui/src/views/home/Home.tsx
@@ -1,12 +1,12 @@
 import {Fragment} from "react";
 import Navbar from "../../components/Navbar";
-
-
+import Hero from "../../components/Hero";
 export default function Home(){
 
   return(
     <Fragment>
       <Navbar></Navbar>
+      <Hero></Hero>
     </Fragment>
   )
 }
